Removing or moving the include files for qtio which may or may not need header files
